move checksum logic back to websource due to circular import
