+++ src/pymovements/dataset/dataset_files.py
@@ -20,9 +20,6 @@
 """Functionality to scan, load and save dataset files."""
 from __future__ import annotations
 
-import errno
-import hashlib
-import os
 from collections.abc import Sequence
 from copy import deepcopy
 from dataclasses import dataclass
@@ -50,35 +47,6 @@ from pymovements.gaze.io import from_ipc
 from pymovements.measure.reading import ReadingMeasures
 from pymovements.stimulus.image import ImageStimulus
 from pymovements.stimulus.text import TextStimulus
-
-
-@dataclass(slots=True, eq=False)
-class ChecksumError(Exception):
-    """Exception raised when a checksum integrity check fails.
-
-    Attributes
-    ----------
-    expected: str
-        Expected checksum.
-    actual: str
-        Actual checksum.
-    path: Path
-        Path of checked file.
-    algorithm: str
-        Name of the checksum algorithm. (default: 'MD5')
-    """
-
-    expected: str
-    actual: str
-    path: Path
-    algorithm: str = 'MD5'
-
-    def __str__(self) -> str:
-        """Get exception message."""
-        return (
-            f"{self.algorithm} checksum mismatch for file '{self.path}'"
-            f": expected '{self.expected}', got '{self.actual}'"
-        )
 
 
 @dataclass
@@ -123,6 +91,7 @@ class DatasetFile:
         if metadata is None:
             metadata = {}
         self.metadata = metadata
+
 
 def scan_dataset(
         definition: DatasetDefinition, paths: DatasetPaths,
